moved argument addition into editor
