+++ .gitignore
@@ -95,3 +95,6 @@ test/Sentry.Unity.Tests/other/**
 samples/unity-of-bugs/*IL2CPPCache/
 samples/unity-of-bugs/*IL2CPPStats/
 *Player.link.log
+
+# Gradle caches generated by the VS Code Gradle extension scanning test fixtures
+.gradle/
